Remove pipes

+++ load_cell/run.sh
@@ -4,18 +4,32 @@ set -eo pipefail
 
 ./unbind.sh
 
+# remove pipes if they exist
+if [ -p pipe0 ]; then
+    rm pipe0
+fi
+if [ -p pipe1 ]; then
+    rm pipe1
+fi
+
+# make pipes
+mkfifo pipe0
+mkfifo pipe1
+
 (
     set -eo pipefail
-    mkfifo pipe0
-    python read_messages.py http://localhost:3000 0 LoadCell0 > pipe0 & READ_PID=$!
+    # start read_messages.py in background
+    python read_messages.py http://localhost:3000 0 LoadCell0 > pipe0 & READ_PID_0=$!
+    # start main and write_records.py in foreground
     ./main 652964 1750 < pipe0 | python write_records.py http://localhost:3000 0 LoadCell0
+    # kill read_messages.py
     kill $READ_PID
-    rm pipe0
 ) & (
     set -eo pipefail
-    mkfifo pipe1
+    # start read_messages.py in background
     python read_messages.py http://localhost:3000 0 LoadCell1 > pipe1 & READ_PID=$!
+    # start main and write_records.py in foreground
     ./main 1076702 1750 < pipe1 | python write_records.py http://localhost:3000 0 LoadCell1
+    # kill read_messages.py
     kill $READ_PID
-    rm pipe1
 )
